Move LineWrapper test to line_wrapper.rs
